mk-flash-layout: improve detection of image type
First, we perform a single call to od and read much more data at once
(1082 bytes). Second, we differenciate uinitrd from ukernel or other
uimages, we distinguish a FAT from just an MBR, and we are now able to
detect an ext2 sitting on top of an MBR since this is permitted.