mk-flash-layout: detect tp-link partitions
TP-Link partitions are a determined by parsing the kernel image which
starts with a 512-byte header. Here we detect the header which must be
on MTD1 exclusively, and if found, we return "tplink" instead of nothing.
The variable TPLINK_ROOTFS_LEN is also fed with the reported rootfs
length, in order to later use it to adjust the apparent rootfs size
(which is otherwise not correct).