mk-flash-layout: support mounting the block device upon detection
Older config-blkdev scripts expect mk-flash-layout from leaving the
file system mounted under /flash when called with -k, and restore-etc
does this at boot if it sees that there is no flash-layout.rc. So the
first attempt to mount the FS fails. Of course, these scripts will
have to be adapted to check that the FS was mounted otherwise try to
do it themselves. Anyway since we rely on the information we pass to
these scripts, let's add support for mounting the block device there
when -k is specified.