mk-flash-layout: correctly detect JFFS2 on NAND devices
Contrary to NOR devices, on NAND devices, the JFFS2 signature is
written in the out-of-band area. So we have to make use of nanddump
to dump the first page and the accompanying OOB data to look for
the signature. We only do that on devices which appear as "empty"
since it's the only situation where this can happen.