mk-flash-layout: consider MMC devices in addition to MTD for /flash
Now that we can detect MMC devices with their respective scores, we can
automatically detect what device to mount based on their respective scores.
For this, we check what we have in decreasing score ordering, always giving
the priority to MTD over MMC. The rationale here is that MTD is (almost)
always soldered on board while MMC sometimes is a removable SD card reader.
The highest scored partition is elected as the software partition. Also,
for compatibility with the old config-blkdev tools, we have to try to mount
the block device if -k is passed.