mount-flash: automatically detect missing /cfg in /flash
There's a problem with jffs2 partitions, they can be detected in the
middle of some junk and still appear fine. When trying to mount them,
they can be virgin but will not cause an error. But /cfg is missing,
so save-etc fails.

Here we implement an "auto-fix" mode, except for JFFS2 where we cannot
safely fix it. In complement we have a "fail-on-error" mode with "-e"
which causes the FS to be unmounted and an error to be reported when
it's impossible to fix it. This way we ensure we never leave a /flash
mounted without /cfg in it when this option is set, so either it's a
jffs2 and we suppose it was not formated, so we unmount it and fail,
or it's another type of filesystem, in which case we try to remount it
R/W and create /cfg in it.

The caller should call it with -e if it knows how to format it in case
of failure.