mk-flash-layout: finish the ubifs implementation
We used to only detect UBI devices but never went as far as validating
them. What's done here is that the UBI volume signature is verified
and the device name is created as ubi<devnum>:<volname>. We also append
"-t ubifs" to the mount option since it's required to parse the device
name. This was validated on a NAND on Armada 370 with kernel 5.4.